// src/modules/subcategory/enums/subcategory-message.enum.ts
export enum SubcategoryMessage {
  CREATED = 'Subcategory successfully created.',
  RETRIEVED_ALL = 'List of all subcategories retrieved successfully.',
  RETRIEVED_ONE = 'Subcategory retrieved successfully.',
  UPDATED = 'Subcategory updated successfully.',
  DELETED = 'Subcategory deleted successfully.',
  NOT_FOUND = 'Subcategory not found.',
  CATEGORY_NOT_FOUND = 'Category not found.',
}